+++ src/blob/core.rs
@@ -1,3 +1,5 @@
+use std::time::SystemTime;
+
 use tokio::time::Instant;
 
 use crate::error::ValidationErrorKind;
@@ -22,7 +24,7 @@ where
     name: FileName,
     file: File,
     current_offset: Arc<Mutex<u64>>,
-    key_type_marker: PhantomData<K>,
+    created_at: SystemTime,
 }
 
 impl<K> Blob<K>
@@ -51,7 +53,7 @@ where
             name,
             file,
             current_offset,
-            key_type_marker: PhantomData,
+            created_at: SystemTime::now(),
         };
         blob.write_header().await?;
         Ok(blob)
@@ -59,6 +61,10 @@ where
 
     pub fn name(&self) -> &FileName {
         &self.name
+    }
+
+    pub(crate) fn created_at(&self) -> SystemTime {
+        self.created_at
     }
 
     async fn write_header(&mut self) -> Result<()> {
@@ -148,13 +154,14 @@ where
         };
         trace!("index initialized");
         let header_size = bincode::serialized_size(&header)?;
+        let created_at = file.created_at()?;
         let mut blob = Self {
             header,
             file,
             name,
             index,
             current_offset: Arc::new(Mutex::new(size)),
-            key_type_marker: PhantomData,
+            created_at,
         };
         trace!("call update index");
         if is_index_corrupted || size as u64 > header_size {