+++ src/storage/observer_worker.rs
@@ -1,5 +1,5 @@
 use super::prelude::*;
-use tokio::{sync::mpsc::Receiver, sync::Semaphore, time::timeout};
+use tokio::{sync::mpsc::Receiver, sync::Semaphore};
 
 pub(crate) struct ObserverWorker<K>
 where
@@ -8,7 +8,6 @@ where
     inner: Inner<K>,
     receiver: Receiver<Msg>,
     dump_sem: Arc<Semaphore>,
-    update_interval: Duration,
     async_oplock: Arc<Mutex<()>>,
 }
 
@@ -22,18 +21,15 @@ where
         dump_sem: Arc<Semaphore>,
         async_oplock: Arc<Mutex<()>>,
     ) -> Self {
-        let update_interval = Duration::from_millis(inner.config.update_interval_ms());
         Self {
             inner,
             receiver,
             dump_sem,
-            update_interval,
             async_oplock,
         }
     }
 
     pub(crate) async fn run(mut self) {
-        debug!("update interval: {:?}", self.update_interval);
         loop {
             if let Err(e) = self.tick().await {
                 warn!("active blob will no longer be updated, shutdown the system");
@@ -45,20 +41,15 @@ where
     }
 
     async fn tick(&mut self) -> Result<()> {
-        match timeout(self.update_interval, self.receiver.recv()).await {
-            Ok(Some(msg)) => {
-                self.process_msg(msg).await?;
-            }
-            Ok(None) => {
+        match self.receiver.recv().await {
+            Some(msg) => self.process_msg(msg).await,
+            None => {
                 return Err(anyhow!(
                     "all observer connected to this worker are dropped, so worker is done"
                         .to_string()
                 ))
             }
-            Err(_) => {}
         }
-        trace!("check active blob");
-        self.try_update().await
     }
 
     async fn process_msg(&mut self, msg: Msg) -> Result<()> {
@@ -71,7 +62,7 @@ where
         }
         match msg.optype {
             OperationType::ForceUpdateActiveBlob => {
-                update_active_blob(self.inner.clone()).await?;
+                update_active_blob(&self.inner).await?;
             }
             OperationType::CloseActiveBlob => {
                 self.inner.close_active_blob().await?;
@@ -87,6 +78,13 @@ where
                     .try_dump_old_blob_indexes(self.dump_sem.clone())
                     .await;
             }
+            OperationType::TryUpdateActiveBlob => {
+                if self.try_update_active_blob().await? {
+                    self.inner
+                        .try_dump_old_blob_indexes(self.dump_sem.clone())
+                        .await;
+                }
+            }
         }
         Ok(())
     }
@@ -99,58 +97,50 @@ where
         }
     }
 
-    async fn try_update(&self) -> Result<()> {
-        trace!("try update active blob");
-        let inner_cloned = self.inner.clone();
-        if let Some(inner) = active_blob_check(inner_cloned).await? {
-            update_active_blob(inner.clone()).await?;
-            inner.try_dump_old_blob_indexes(self.dump_sem.clone()).await;
+    async fn try_update_active_blob(&self) -> Result<bool> {
+        let config_max_size = self
+            .inner
+            .config
+            .max_blob_size()
+            .ok_or_else(|| Error::from(ErrorKind::Uninitialized))?;
+        let config_max_count = self
+            .inner
+            .config
+            .max_data_in_blob()
+            .ok_or_else(|| Error::from(ErrorKind::Uninitialized))?;
+
+        {
+            let read = self.inner.safe.read().await;
+            let active_blob = read.active_blob.as_ref();
+            if let Some(active_blob) = active_blob {
+                if active_blob.file_size() < config_max_size
+                    && (active_blob.records_count() as u64) < config_max_count
+                {
+                    return Ok(false);
+                }
+            }
         }
-        Ok(())
+
+        let mut write = self.inner.safe.write().await;
+        let active_blob = write.active_blob.as_ref();
+        if let Some(active_blob) = active_blob {
+            if active_blob.file_size() >= config_max_size
+                || active_blob.records_count() as u64 >= config_max_count
+            {
+                let new_active = get_new_active_blob(&self.inner).await?;
+                write.replace_active_blob(new_active).await?;
+                return Ok(true);
+            }
+        }
+        Ok(false)
     }
 }
 
-async fn active_blob_check<K>(inner: Inner<K>) -> Result<Option<Inner<K>>>
+async fn update_active_blob<K>(inner: &Inner<K>) -> Result<()>
 where
     for<'a> K: Key<'a> + 'static,
 {
-    let (active_size, active_count) = {
-        trace!("await for lock");
-        let safe_locked = inner.safe.read().await;
-        trace!("lock acquired");
-        if let Some(active_blob) = safe_locked.active_blob.as_ref() {
-            (active_blob.file_size(), active_blob.records_count() as u64)
-        } else {
-            // if active blob doesn't exists, it doesn't need to be updated
-            return Ok(None);
-        }
-    };
-    trace!("lock released");
-    let config_max_size = inner
-        .config
-        .max_blob_size()
-        .ok_or_else(|| Error::from(ErrorKind::Uninitialized))?;
-    let config_max_count = inner
-        .config
-        .max_data_in_blob()
-        .ok_or_else(|| Error::from(ErrorKind::Uninitialized))?;
-    if active_size as u64 > config_max_size || active_count >= config_max_count {
-        Ok(Some(inner))
-    } else {
-        Ok(None)
-    }
-}
-
-async fn update_active_blob<K>(inner: Inner<K>) -> Result<()>
-where
-    for<'a> K: Key<'a> + 'static,
-{
-    let next_name = inner.next_blob_name()?;
-    // Opening a new blob may take a while
-    trace!("obtaining new active blob");
-    let new_active = Blob::open_new(next_name, inner.ioring, inner.config.index())
-        .await?
-        .boxed();
+    let new_active = get_new_active_blob(inner).await?;
     inner
         .safe
         .write()
@@ -158,4 +148,16 @@ where
         .replace_active_blob(new_active)
         .await?;
     Ok(())
+}
+
+async fn get_new_active_blob<K>(inner: &Inner<K>) -> Result<Box<Blob<K>>>
+where
+    for<'a> K: Key<'a> + 'static,
+{
+    let next_name = inner.next_blob_name()?;
+    trace!("obtaining new active blob");
+    let new_active = Blob::open_new(next_name, inner.ioring.clone(), inner.config.index())
+        .await?
+        .boxed();
+    Ok(new_active)
 }