+++ CHANGELOG.md
@@ -7,7 +7,7 @@ Pearl changelog
 
 
 #### Changed
-
+- Remove active checking for blob update (#159)
 
 #### Fixed
 - corrupted blob now should be saved in case of 'unexpected eof' error